order-service/pyproject.toml
[tool.poetry]
name = "order-service"
version = "0.1.0"
description = "Order Service for Real-Time Order Processing System"
authors = ["Jatin Yadav <jatinyadav154@gmail.com>"]
packages = [{include = "app"}]  

[tool.poetry.dependencies]
python = ">=3.13"  
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
aiokafka = "^0.12.0"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
python-dotenv = "^1.1.0"
greenlet = "^3.1.1"

[tool.poe.tasks]
runserver = "uvicorn app.main:app --reload"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.33.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
